requirements.txt
Flask
Flask-PyMongo
Flask-Bcrypt
Flask-JWT-Extended
pytest